Backend/requirements.txt
Flask==2.0.1
flask-cors==3.0.10
pymongo==3.11.4
mysql-connector-python==8.0.25
Werkzeug==2.0.1